Add default category, category validations, and question body to routes
A default category is seeded, but is dependent on the front-end for
defaulting all newly created questions to be included in this category.

Categories now need to have a category name as a bare minimum - an
nameless category is quite useless.

The question routes now include the posting and putting of a question
body q_body.