General housekeeping
Removed useless comments from controller and made a note in services
that sharing the controller among 2 directives is calling the controller
twice. Since the initialisation code is not wrapped in a function that
will run only once, it is called twice, making the code call the backend
API twice as many times as necessary.